build: make the terminal library compile under MSVC, and prove the char32_t work
The char32_t change was verified on Linux, where the bug it fixes cannot occur
-- wchar_t is already 32 bits there. This makes the evidence real by building
and running the tests on the platform where wchar_t is 16 bits.

Three things were in the way, none of them large:

  * __attribute((unused)) at 25 sites. MSVC does not merely warn about it, it
    is a syntax error. Replaced with MOSH_UNUSED from the new
    src/util/compat.h, which is also where the platform layer's other small
    spellings will go. The one remaining instance is inside
    #if defined(OCB_TEST_PROGRAM), which neither build system defines.

  * <unistd.h> in terminal.cc and terminalfunctions.cc. Neither file uses a
    single symbol from it -- checked, not assumed.

  * moshutil mixes pure computation with POSIX. It is now split:
    moshutil_portable (utf8, uniwidth) builds anywhere, and moshutil (select,
    pty_compat, swrite, timestamp, locale_utils) is POSIX-only until the
    platform layer exists. src/terminal turns out to depend on exactly the two
    portable headers, so moshterminal now builds on Windows once
    terminaldisplayinit.cc -- the terminfo probe, replaced later by a
    compiled-in table -- is excluded there.

Results under MSVC 14.44:

    unicode-roundtrip   0 failures
    uniwidth            table hash 0x1313ab78fa2e8001, 0 failures
    utf8                3210197 checks, 0 failures

The round-trip test carries U+1F600, U+20000, U+1D11E and U+10FFFD through the
parser, the emulator and the framebuffer. Every one of those would have been
truncated to a lone surrogate before this change, and Cell::append() would have
called contents.insert() with a length of (size_t)-1.

The uniwidth hash is the same value the Linux build produces. That is the
cross-check that matters: mosh's diff protocol requires the client and the
server to compute identical widths, and a disagreement is silent -- it
misplaces everything after the first divergent wide character rather than
reporting anything.

Linux is unaffected: 35 tests, 33 pass and 2 expected failures, unchanged.

+++ src/tests/utf8.cc
@@ -89,8 +89,8 @@ static void test_roundtrip( void )
     size_t got = Util::utf8_decode( &back, buf, n );
     if ( got != ( wc == 0 ? 0u : n ) || back != wc ) {
       char d[128];
-      snprintf( d, sizeof d, "U+%04X encoded as %s decoded to U+%04X (ret %zd)", (unsigned)wc,
-                hex( buf, n ).c_str(), (unsigned)back, (ssize_t)got );
+      snprintf( d, sizeof d, "U+%04X encoded as %s decoded to U+%04X (ret %lld)", (unsigned)wc,
+                hex( buf, n ).c_str(), (unsigned)back, (long long)got );
       fail( "round trip", d );
     }
     /* Every proper prefix must report INCOMPLETE, never INVALID: that is what
@@ -189,8 +189,8 @@ static void compare_with_libc( const char* bytes, size_t n )
 
   if ( mine_bad != their_bad || mine_short != their_short ) {
     char d[192];
-    snprintf( d, sizeof d, "%s -- mine %zd, libc %zd", hex( bytes, n ).c_str(), (ssize_t)mine_ret,
-              (ssize_t)their_ret );
+    snprintf( d, sizeof d, "%s -- mine %lld, libc %lld", hex( bytes, n ).c_str(), (long long)mine_ret,
+              (long long)their_ret );
     fail( "disagreement on validity", d );
     return;
   }
